Add updatedAt to copySubmission method
- Updates @Transactional and @Service calls to use the Spring Boot library versions instead of Jakarta versions as well
- Moves copySubmission and copyMap methods from SubmissionUtils to Submission object and deletes SubmissionUtils

[#185414835]
