+++ src/main/java/formflow/library/data/Submission.java
@@ -173,6 +173,53 @@ public class Submission {
     inputData.remove(inputName + AddressParts.ZIPCODE + UnvalidatedField.VALIDATED);
   }
 
+  /**
+   * Create a deep copy of the given submission.
+   *
+   * @param origSubmission given submission to copy
+   * @return deep copy of origSubmission
+   */
+  public static Submission copySubmission(Submission origSubmission) {
+    Submission newSubmission = new Submission();
+    newSubmission.setUrlParams(new HashMap<>(origSubmission.getUrlParams()));
+
+    newSubmission.setFlow(origSubmission.getFlow());
+    newSubmission.setCreatedAt(origSubmission.getCreatedAt());
+    newSubmission.setUpdatedAt(origSubmission.getUpdatedAt());
+    newSubmission.setSubmittedAt(origSubmission.getSubmittedAt());
+    newSubmission.setId(origSubmission.getId());
+
+    // deep copy the subflows and any lists
+    newSubmission.setInputData(copyMap(origSubmission.getInputData()));
+    return newSubmission;
+  }
+
+
+  private static Map<String, Object> copyMap(Map<String, Object> origMap) {
+    Map<String, Object> result = new HashMap<>();
+    for (Map.Entry<String, Object> entry : origMap.entrySet()) {
+      if (entry.getValue() instanceof List) {
+        List data = (List) entry.getValue();
+        if (data.size() == 0) {
+          result.put(entry.getKey(), new ArrayList<>());
+        } else if (data.get(0) instanceof String) {
+          result.put(entry.getKey(), new ArrayList<>(data));
+        } else if (data.get(0) instanceof Map) {
+          List<Map> newList = new ArrayList<>();
+          for (Map element : (List<Map>) data) {
+            newList.add(copyMap(element));
+          }
+          result.put(entry.getKey(), newList);
+        } else {
+          result.put(entry.getKey(), entry.getValue());
+        }
+      } else {
+        result.put(entry.getKey(), entry.getValue());
+      }
+    }
+    return result;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
