+++ src/main/java/formflow/library/data/SubmissionRepositoryService.java
@@ -1,13 +1,12 @@
 package formflow.library.data;
 
 import jakarta.servlet.http.HttpSession;
-import jakarta.transaction.Transactional;
-import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Service to retrieve and store Submission objects in the database.
